requirements.txt
channel==3.0.2
django==2.2.16
django-cors-headers==3.5.0
django-redis==4.12.1
psycopg2-binary==2.8.6
urllib3==1.25.8
uWSGI==2.0.19.1
redis==3.5.3
requests==2.22.0